feat(#10): post 리스트 조회 및 검색조회

+++ demo/src/main/java/com/union/demo/entity/Role.java
@@ -13,7 +13,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="role_id")
-    private Long roleId;
+    private Integer roleId;
 
     @Column(name="field_id", nullable = false)
     private Long fieldId;
